refactor(ui): React 유틸리티 모듈 분리
비-React 헬퍼와 수명주기 reducer를 전용 모듈로 옮겨 Fast Refresh 경계를 명확히 하고 기존 UI 계약을 유지한다.
