+++ frontend/src/app/external-link-controller.tsx
@@ -2,23 +2,10 @@ import {CircleAlert, X} from 'lucide-react';
 import {useMemo, useState, type PropsWithChildren} from 'react';
 
 import {Button} from '@/components/ui/button';
-import {ExternalLinkRuntimeProvider, type ExternalUrlOpener} from '@/components/ui/external-link';
+import {ExternalLinkRuntimeProvider} from '@/components/ui/external-link';
 import type {PlatformAdapter} from '@/platform/contracts';
 
-function hasExternalLinkOpener(value: unknown): value is {open: ExternalUrlOpener} {
-    return (
-        typeof value === 'object' &&
-        value !== null &&
-        'open' in value &&
-        typeof value.open === 'function'
-    );
-}
-
-export function desktopExternalOpener(platform: PlatformAdapter): ExternalUrlOpener | undefined {
-    if (platform.kind !== 'desktop' || !('externalLinks' in platform)) return undefined;
-    const externalLinks = platform.externalLinks;
-    return hasExternalLinkOpener(externalLinks) ? (url) => externalLinks.open(url) : undefined;
-}
+import {desktopExternalOpener} from './external-link-runtime';
 
 export function DashboardExternalLinkController({
     children,
