+++ frontend/src/components/ui/external-link.tsx
@@ -7,6 +7,7 @@ import {
     type ReactElement,
 } from 'react';
 
+import {openExternalLinkFromClick} from './external-link-click';
 import {normalizeExternalUrl} from './external-link-policy';
 
 export type ExternalUrlOpener = (url: string) => Promise<void>;
@@ -33,21 +34,6 @@ export interface ExternalLinkProps extends Omit<ComponentProps<'a'>, 'href' | 't
     href: string;
     openExternally?: ExternalUrlOpener;
     onOpenError?: (error: unknown) => void;
-}
-
-export async function openExternalLinkFromClick(
-    event: Pick<Event, 'defaultPrevented' | 'preventDefault'>,
-    href: string,
-    openExternally: ExternalUrlOpener | undefined,
-    onOpenError: ((error: unknown) => void) | undefined,
-): Promise<void> {
-    if (event.defaultPrevented || !openExternally) return;
-    event.preventDefault();
-    try {
-        await openExternally(href);
-    } catch (error) {
-        onOpenError?.(error);
-    }
 }
 
 export function ExternalLink({
